feat(chain_monitor)!: Split block process from confirmation report
BREAKING CHANGE: Consumers have to call `periodic_chain_monitor`
periodically so that `dlc_manager::Manager` and `SubChannelManager`
can learn about transactions being confirmed. Furthermore, the
`Writeable` and `Readable` traits of `ChainMonitor` have changed, so
we are incompatible with `ChainMonitor`s persisted before this patch.

There are two consumers of the `ChainMonitor` API:
`dlc_manager::Manager` and `SubChannelManager`. Both of them are
interested in knowing when certain transactions reach a confirmation
target (for now a single confirmation).

Before this patch, the only way to learn about confirmed transactions
was to call `process_block`. When using both managers (when using
LN-DLC channels) this presented two options:

1. Let each manager provide their own block.

2. Have the `SubChannelManager` process the block throught the
`ChainMonitor` and reuse the same result to let the
`dlc_manager::Manager` process it.

The first option led to bugs fixed in
https://github.com/p2pderivatives/rust-dlc/commit/4636ad7e7bc15ba524e7be66b6d5a0a86f8f8999.
The second option was what was implemented before this patch and it
resulted in coupling between the two managers.

With this patch we now separate the processing of new blocks from the
managers learning about confirmed transactions. This does mean that
library consumers will have to periodically call
`periodic_chain_monitor` to update the state of `ChainMonitor`. It's
less convenient, but it's not different from similar projects such as
`rust-lightning`.